+++ src/main/java/RhemaApp/Rhema/entity/Position.java
@@ -1,4 +1,4 @@
-package entity;
+package RhemaApp.Rhema.entity;
 
 public class Position {
     private int minutes;
