+++ src/main/java/RhemaApp/Rhema/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package RhemaApp.Rhema.repository;
 
-import entity.User;
+import RhemaApp.Rhema.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
